Move header `collector.hpp` into namespace `kernel` from `memory`
