+++ src/kernel/pair.cpp
@@ -16,13 +16,10 @@
 
 #include <meevax/kernel/list.hpp>
 
-namespace meevax::inline memory
-{
-  template struct collector<pair, bool, small_integer, float, character, instruction>;
-}
-
 namespace meevax::inline kernel
 {
+  template struct collector<pair, bool, small_integer, float, character, instruction>;
+
   let unit = nullptr;
 
   struct datum_labels
